Mise en place de zod pour les validators
