+++ back/src/api/declaration/getDeclarationWithPublicId/validator.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod';
-import { buildValidationMiddleware } from '../../../core/middlewares/zodValidation.middleware';
+import { buildValidationMiddleware } from '../../../core/middlewares';
 
 export const getOneDeclarationWithPublicIdValidator = z.object({
   params: z.object({
