+++ back/src/api/product/putSearchHistory/validator.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod';
-import { buildValidationMiddleware } from '../../../core/middlewares/zodValidation.middleware';
+import { buildValidationMiddleware } from '../../../core/middlewares';
 
 export const putSearchProductHistoryValidator = z.object({
   body: z.object({
