+++ back/tests/api/authentication/register/validator.test.ts
@@ -1,8 +1,7 @@
 import { faker } from '@faker-js/faker';
 import { registerValidator } from '../../../../src/api/authentication/register/validator';
 import { config } from '../../../../src/loader/config';
-import { zodValidatorHelper } from '../../../../src/core/testHelpers/zodValidator.helper';
-
+import { validatorHelper } from '../../../../src/core/testHelpers';
 describe('register validator', () => {
   const validator = registerValidator;
   const validData = {
@@ -11,7 +10,7 @@ describe('register validator', () => {
       password: 'Password95*',
     },
   };
-  const { isValid } = zodValidatorHelper(validator);
+  const { isValid } = validatorHelper(validator);
 
   it('should validate proper data - douane email', () => {
     expect(isValid(validData)).toBeTruthy();
